fix some click/router bugs

+++ SimilarWordWeb/ClientApp/src/components/WordEdit.js
@@ -2,10 +2,21 @@
 import { Form, FormControl, FormGroup, ControlLabel, Col, Button, Label } from 'react-bootstrap';
 import { ShowDictLink, GetTokenHeader } from './CommTools';
 
-const EditWord = ({ name, location }) => {
+const EditWord = (props) => {
     // Declare a new state variable, which we'll call "count"
-    const [word, setWord] = useState(location.state.word);
+    const [word, setWord] = useState(props.location && props.location.state ? props.location.state.word : {});
     const [result, setResult] = useState('');
+
+    console.log("EditWord start");
+    console.log(props);
+    if (!props.location ||!props.location.state ) return (
+        <Fragment>
+            <span>Sorry, No word to edit</span>
+            < button onClick={() => props.history.goBack()}> BACK</button >
+        </Fragment>
+    )
+
+    let { name, location, ...restProps } = props;
 
     const CallUpdateApi = (e) => {
         //alert("in CallUpdateApi");
@@ -28,6 +39,7 @@ const EditWord = ({ name, location }) => {
     const CallDeleteApi = (e) => {
         //alert("in CallUpdateApi");
         e.preventDefault();
+        if (!window.confirm('Are you sure to delete ' + word.name)) return;
         fetch('api/Words/' + word.name, {
             method: 'DELETE',
             body: JSON.stringify(word),
@@ -66,11 +78,12 @@ const EditWord = ({ name, location }) => {
         setResult('');
     }  
 
-    console.log(name);
+    console.log('WordEdit:'+word.name);
     console.log(location.state.word);
+    console.log(restProps);
+
     return (
-        <Fragment>
-        <Form horizontal onSubmit={this.CallUpdateApi} >
+            <Form horizontal onSubmit={this.CallUpdateApi} >
             <h3 >Edit Word: {word.name}</h3>
                 <FormField type='text' label="Word Name" name="name" onChangeHandle={handleChange}  val={word.name} ></FormField>
                 <FormField type='text' label="Pronounciation" name="pronounciation" onChangeHandle={handleChange}  val={word.pronounciation} ></FormField>
@@ -79,18 +92,17 @@ const EditWord = ({ name, location }) => {
                 <FormField type='textarea' label="Meaning" name="meaningShort" onChangeHandle={handleChange}  val={word.meaningShort} ></FormField>
                 {/* <FormField type='textarea' label="Meaning Long" name="meaningLong" onChangeHandle={handleChange} val={word.meaningLong} ></FormField> */}
                 {/* <FormField type='text' label="Meaning Other" name="meaningOther" onChangeHandle={handleChange} val={word.meaningOther} ></FormField>  */}
-            <ShowDictLink name={word.name} > </ShowDictLink><br/>
+                <ShowDictLink name={word.name} > </ShowDictLink><br />
                 <label>{result}</label>
             <FormGroup>
                 <Col smOffset={2} sm={8}>
                     <button className="primary" onClick={CallUpdateApi}>Update</button>
-                    {' '} <button className="primary" onClick={CallDeleteApi}>Delete</button>
+                        {' '} <button className="primary" onClick={CallDeleteApi}>Delete</button>
+                        {' '} <button onClick={() => restProps.history.goBack()}>BACK</button>
                 </Col>
             </FormGroup>
             </Form>
 
-
-        </Fragment>
     );
